refactor: seperate low-level api from routing
This seperates several low-level APIs from the high-level routing:
- TusUploadFile (file.py): class for the upload file itself
- TusUploadInfo (info.py): class for the file's .info param file
- TusUploadParams (params.py): pydantic model for upload params

Additionally, some helper functions are moved to a request.py file, and
routes are seperated into distinct files for the core protocols and its
extensions. Finally, deprecated routes are removed.

The motivation is to improve understandibility of the codebase, and
facilitate maintainance going forward.